+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQ.java
@@ -6,7 +6,7 @@ import com.rabbitmq.client.AMQP;
  * @author neo
  */
 public interface RabbitMQ {
-    void publish(String exchange, String routingKey, String message, AMQP.BasicProperties properties);
+    void publish(String exchange, String routingKey, byte[] message, AMQP.BasicProperties properties);
 
     RabbitMQConsumer consumer(String queue, int prefetchCount);
 }