+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQPublisher.java
@@ -51,7 +51,7 @@ public class RabbitMQPublisher<T> implements MessagePublisher<T> {
 
         linkContext(builder, headers);
 
-        rabbitMQ.publish(exchange, routingKey, JSON.toJSON(message), builder.build());
+        rabbitMQ.publish(exchange, routingKey, JSON.toJSONBytes(message), builder.build());
     }
 
     private void linkContext(AMQP.BasicProperties.Builder builder, Map<String, Object> headers) {