+++ log-processor/src/main/java/core/log/queue/RabbitMQMessage.java
@@ -3,14 +3,12 @@ package core.log.queue;
 /**
  * @author neo
  */
-public class RabbitMQMessage<T> {
+public class RabbitMQMessage {
     public final long deliveryTag;
-    public final T body;
-    public final int messageSize;
+    public final byte[] body;
 
-    public RabbitMQMessage(long deliveryTag, T body, int messageSize) {
+    public RabbitMQMessage(long deliveryTag, byte[] body) {
         this.deliveryTag = deliveryTag;
         this.body = body;
-        this.messageSize = messageSize;
     }
 }